fix review point
