+++ AnimalFriends/Assets/Graphics/Kari.meta
