+++ AnimalFriends/Assets/Graphics/Kari/golem_move.png
