+++ AnimalFriends/Assets/Graphics/Kari/golem_move_0.controller.meta
