--- AnimalFriends/Assets/Scripts/CameraControl.cs
@@ -1,24 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class CameraControl : MonoBehaviour {
-
-	private GameObject player = null;
-	private Vector3 offset = Vector3.zero;
-
-	void Start()
-	{
-		player = GameObject.FindGameObjectWithTag("Player");
-		offset = transform.position - player.transform.position;
-	}
-
-	void LateUpdate()
-	{
-		Vector3 newPosition = transform.position;
-		newPosition.x = player.transform.position.x + offset.x;
-		newPosition.y = player.transform.position.y + offset.y;
-		newPosition.z = player.transform.position.z + offset.z;
-		transform.position = newPosition;
-	}
-}
