+++ AnimalFriends/Assets/Scripts/PlayerController.cs.meta
